Hook back the CacheControlInterceptor and remove the unused StripMustRevalidateResponseInterceptor.
