+++ app/src/main/java/org/wikipedia/dataclient/okhttp/OkHttpConnectionFactory.kt
@@ -26,7 +26,7 @@ object OkHttpConnectionFactory {
                 .addInterceptor(HttpLoggingInterceptor().setLevel(Prefs.getRetrofitLogLevel()))
                 .addInterceptor(UnsuccessfulResponseInterceptor())
                 .addInterceptor(StatusResponseInterceptor(RbSwitch.INSTANCE))
-                .addNetworkInterceptor(StripMustRevalidateResponseInterceptor())
+                .addNetworkInterceptor(CacheControlInterceptor())
                 .addInterceptor(CommonHeaderRequestInterceptor())
                 .addInterceptor(DefaultMaxStaleRequestInterceptor())
                 .addInterceptor(OfflineCacheInterceptor(SAVE_CACHE))
